-Houses Link Impl with Category Table Impl (#151)

+++ src/main/kotlin/com/realtor/plugins/dao/ContactDao.kt
@@ -10,7 +10,7 @@ interface ContactDao {
     ): Contact?
 
     suspend fun getAllContacts(): List<Contact>?
-    suspend fun getContactById(id: Int): Contact?
-    suspend fun deleteContactById(id: Int): Int?
-    suspend fun updateContactById(id: Int,name: String,email: String,message: String): Int?
+    suspend fun getContactById(id: Long): Contact?
+    suspend fun deleteContactById(id: Long): Int?
+    suspend fun updateContactById(id: Long,name: String,email: String,message: String): Int?
 }
